app/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
Django==2.0.7
djangorestframework==3.8.2
kombu==4.2.1
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
vine==1.1.4
